mysqlclass/requirements.txt
Flask==3.0.0
Flask_SQLAlchemy==3.1.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
